Added ProfileCardTests and locator
