+++ jdi-light-vuetify/src/main/java/com/epam/jdi/light/vuetify/elements/complex/data/ProfileCardLocator.java
@@ -1,0 +1,8 @@
+package com.epam.jdi.light.vuetify.elements.complex.data;
+
+public class ProfileCardLocator {
+
+    public static final String PROFILE_CARD_LOCATOR = "//div[@id = 'profileCard']";
+
+}
+
